video-engine/pyproject.toml
[tool.poetry]
name = "video-engine"
version = "0.1.0"
description = "Standalone video processing engine"
authors = ["Pathway Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = "^0.22.0"
yt-dlp = "^2025.12.8"
pyyaml = "^6.0.3"
pillow = "11.1.0"
cairosvg = "^2.8.2"
google-generativeai = "^0.3.2"
elevenlabs = "^0.2.27"
python-dotenv = "^1.0.0"
nicegui = "^3.4.1"
openai-whisper = "*"
setuptools-rust = "*"

[tool.poetry.scripts]
init = "app.scripts.init_engine:main"
gen-assets = "app.scripts.generate_vfx:main"
list-models = "app.scripts.list_models:list_elevenlabs_resources"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
